requirements.txt
pandas~=1.0.3
matplotlib~=3.2.1
pytrends~=4.7.2
numpy~=1.17.1
phpserialize~=1.3
PyYAML~=5.3.1
PyMySQL~=0.9.3
requests~=2.22.0